Extract JSON serialization to new namespace
