+++ FortnoxAPILibrary/Entities/Contract.cs
@@ -1,5 +1,5 @@
 using System.Collections.Generic;
-using System.ComponentModel;
+using FortnoxAPILibrary.Serialization;
 using Newtonsoft.Json;
 
 // ReSharper disable UnusedMember.Global
