+++ FortnoxAPILibrary/Entities/Currency.cs
@@ -1,5 +1,4 @@
-using System;
-using System.ComponentModel;
+using FortnoxAPILibrary.Serialization;
 using Newtonsoft.Json;
 
 // ReSharper disable UnusedMember.Global
