+++ FortnoxAPILibrary/Entities/FortnoxError.cs
@@ -1,4 +1,5 @@
-﻿using Newtonsoft.Json;
+﻿using FortnoxAPILibrary.Serialization;
+using Newtonsoft.Json;
 
 // ReSharper disable UnusedMember.Global
