+++ FortnoxAPILibrary/Entities/LockedPeriod.cs
@@ -1,3 +1,4 @@
+using FortnoxAPILibrary.Serialization;
 using Newtonsoft.Json;
 
 // ReSharper disable UnusedMember.Global
