+++ FortnoxAPILibrary/Entities/SieSummary.cs
@@ -1,5 +1,5 @@
-using System;
 using System.Collections.Generic;
+using FortnoxAPILibrary.Serialization;
 using Newtonsoft.Json;
 
 // ReSharper disable UnusedMember.Global
