+++ FortnoxAPILibrary/Entities/SupplierInvoiceExternalURLConnections.cs
@@ -1,4 +1,4 @@
-using System.ComponentModel;
+using FortnoxAPILibrary.Serialization;
 using Newtonsoft.Json;
 
 // ReSharper disable UnusedMember.Global
