+++ FortnoxAPILibrary/Serialization/EntityAttribute.cs
@@ -1,0 +1,10 @@
+﻿using System;
+
+namespace FortnoxAPILibrary.Serialization
+{
+    internal class EntityAttribute : Attribute
+    {
+        public string SingularName { get; set; }
+        public string PluralName { get; set; }
+    }
+}
