+++ FortnoxAPILibrary/Serialization/GenericPropertyNameAttribute.cs
@@ -1,0 +1,8 @@
+﻿using System;
+
+namespace FortnoxAPILibrary.Serialization
+{
+    internal class GenericPropertyNameAttribute : Attribute
+    {
+    }
+}
