+++ FortnoxAPILibrary/Serialization/ReadOnlyAttribute.cs
@@ -1,0 +1,8 @@
+﻿using System;
+
+namespace FortnoxAPILibrary.Serialization
+{
+    public class ReadOnlyAttribute : Attribute
+    {
+    }
+}
